Changes: - Modified the amount field to double & Updated the unit test.
ToDo
- Read Column Names without explicitly hardcoding the column names to the FileItemReader.
- Link all transactions to the associated accounts.
- Read from Directory without hardcoding the absolute path.

+++ src/test/java/io/vultr/cld/paycons/persistence/repo/TxRepoTest.java
@@ -7,7 +7,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.test.context.ActiveProfiles;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.UUID;
@@ -30,7 +29,7 @@ class TxRepoTest {
     @Test
     void givenTxRepo_whenSavingNewEntry_thenReturnSavedTx() {
 
-        final Tx tx = Tx.builder().ref(ref).date(LocalDate.now()).time(LocalTime.now()).descr("IFT Withdrawal").type("C").amount(BigDecimal.valueOf(89.99)).build();
+        final Tx tx = Tx.builder().ref(ref).date(LocalDate.now()).time(LocalTime.now()).descr("IFT Withdrawal").type("C").amount(89.99D).build();
 
         txRepo.save(tx);
         final Tx txResp = txRepo.findById(1L).get();
